made new tests file
